+++ docs/changelog.md
@@ -22,6 +22,7 @@ category: Administration
 * docs: include additional fields in feature overview (#128)
 * feat: show group icon on issue detail page to start or join session
 * docs: rename and shorten pages for configuration and installation
+* feat: allow the estimate field to be overwritten per project (#103)
 
 ### [20.10.0] - 2020-10-08